moved rangeF and rangeD to same location as range function. Oops.
